requirements.txt
requests
datasets
pandas
python-dotenv 
openai 
matplotlib 
seaborn 
tqdm
typing
openai
anthropic
google-generativeai
google-cloud-aiplatform
google
scipy
scikit-learn
